Latest sorting out. Everything seems to agree now. Tested for me.

+++ users/ericgebhart/extensions/keymap_combo.h
@@ -126,7 +126,7 @@ void process_combo_event(uint16_t combo_index, bool pressed) {
 #undef COMBO_REF_LAYER
 #define COMBO_REF_LAYER REF_LAYER
 
-uint16_t combo_ref_from_layer(uint16_t layer){
+uint8_t combo_ref_from_layer(uint8_t layer){
   switch (biton32(layer_state)){
 #include "combos.def"
